Fixes
Should now correctly identify and display DevFee, as well as identify
miners with or without rigNames supplied (by ip:port, -eworker, or
.rigName)
Some groundwork for -eworker pools as well. It connects to a Ubiq pool
and seems to work. Does not deal with the DevFee as of yet, as the
DevFee is going through -allcoins 1, which means it mines on a different
pool for the DevFee. Jokes on him, though, my Ubiq miner only has 2GB
cards in it. >:)