Initial Monero protocol
It works. I don't understand a lot of the protocol, and I don't see how
you identify your rigs, so I use the endPoint as the name.
I can detect submitted shares, but all the server returns is an OK,
without reference to whether that OK is about a share or not.
But it does replace the DevFee wallet.