Implemented new _consoleQueue
and changed existing _consoleQueue to _webConsoleQueue.
I noticed the console is slowing things down when there are many
connections and the CPU can't easily keep up with the console writes, so
I wanted to offload some of this to a separate Task, and have it do them
one at a time as they come in. This may result in a little more RAM
usage, but should help with performance in the long run.
Must now specift bypassQueue when calling LogToConsole in order to log
directly, otherwise it's added to the queue. ConsoleBlocks are unchanged
currently, hopefully it isn't an issue and will keep some text together.
Will test more
Added endPoints to webConsole as well.