Failover pool support
Added a first draft of failover support.
After a few connection failure attempts, the next pool in the list is used for new connections.
Virtually unlimited number of failover pools, in a round-robin format. List should loop back to beginning after reaching the end.

Does not work for lack of accepted shares or any other scenario as of now.

May not work as expected. May not work at all. I've only briefly tested it.